move all Bebop code into one header and add bebop_test folder
