+++ bebop_test/bebop_test.cc
@@ -1,0 +1,6 @@
+#include "robots/bebop.h"
+
+int main()
+{
+    
+}
